refactor: construct classes dynamically
